devcon: add TTY abstraction
For each terminal window we create in devcon, we need a PTY for userspace
to talk to us. As we cannot use the unix-PTYs from within the kernel
(well, we could, but that would become really ugly really fast), we
implement our own TTY driver, just like the VT layer does.

Instead of using static TTY allocations (like VTs do), we implement a
fully dynamic scheme. This allows us to operate much more like PTYs and
use TTYs dynamically whenever we require one.

The I/O callbacks are not hooked up, yet. This will be done in followup
patches.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>